Add snapshot tests for obstacles, move mock environment
