+++ game_frontend/src/babylon/entities/obstacle.ts
